Allow users to specify a custom JSON Request Template for State Machine API Events
Add `JsonRequestTemplate` as a permissible property for the Api method event source for SAM State Machines.  This allows a user to specify an alternate application/json request template.  The primary use case is to make using the documented `.replaceAll("\\'", "'")` variation possible without losing the simplicity of the implicit APIs but it also enables arbitrary request templates.

Update `_generate_request_template(self, resource)` to use a supplied value if `JsonRequestTemplate` is supplied as a property of the Api event for the State Machine or use the current default if not.

Add a custom encoder for `json.dumps` to use to encode the request template.  The default implementation makes it impossible to specify the `replaceAll` variant.  This implementation inserts the string 'as is' which is essentially invalid JSON but CloudFormation tolerates it and it deploys and works as it should.  The custom encoder removes any trailing newlines that could appear in the input if the user uses a pipe symbol to write the RequestTemplate in the SAM template.